Initial commit of changelog
This initial version supports generating a HTML changelog from GitHub
Issues that are referenced in commits that are part of a Jenkins job.

Supports two commands:
- print_html: prints the changelog of an entire Jenkins job’s history
as HTML
- notify: emails the changelog of a single Jenkins build

Upon first launch, it will prompt for any necessary credentials and
configuration, and write it to `~/.changelog.conf`. As long as that
configuration file exists, it will load the configuration from there
and not prompt you.